% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cribbR.R
\name{add_snippet_edit_shortcut}
\alias{add_snippet_edit_shortcut}
\title{Function adding a snippet to easily editing the snippet file.}
\usage{
add_snippet_edit_shortcut(snippet_path)
}
\arguments{
\item{snippet_path}{The path pointing to your snippet file. Returned by the function 'locate_snippet_file()'}
}
\value{
Checks if a edit_snippet snippet exists. If not, asks if you want to add it automatically or manually.
}
\description{
Function adding a snippet to easily editing the snippet file.
}
