% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cribbR.R
\name{determine_system}
\alias{determine_system}
\title{Determine operating system}
\usage{
determine_system(os_mode = NULL)
}
\arguments{
\item{os_mode}{Manual override of which OS to use. Probably only useful for developer.}
}
\value{
What operation system is run
}
\description{
Determine operating system
}
