% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cribbR.R
\name{let_there_be_snippets}
\alias{let_there_be_snippets}
\title{Wrapper function to locate snippet file and, if it exists, offers to add an 'edit snippet'-snippet.}
\usage{
let_there_be_snippets(
  be_my_guess = NULL,
  win_user_name = NULL,
  rstudio_path = NULL,
  snippet_path = NULL
)
}
\arguments{
\item{be_my_guess}{Looks for user folders on Windows if TRUE. Set to TRUE if no arguments are specified.}

\item{win_user_name}{The name of window user folder to use.}

\item{rstudio_path}{Path to windows folder containing the R-studio installation ("c:/user/AppData/Roaming/Rstudio/")}

\item{snippet_path}{Path pointing to your snippet file ("r.snippets")}
}
\value{
Checks if the snippet file exists. If it does, checks if a edit_snippet snippet exists. If not, asks if you want to add it automatically or manually.
}
\description{
Wrapper function to locate snippet file and, if it exists, offers to add an 'edit snippet'-snippet.
}
