% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cribbR.R
\name{robust_snippet_find}
\alias{robust_snippet_find}
\title{A potentially more solid version to find the r.snippet}
\usage{
robust_snippet_find()
}
\value{
The path to the snippet-file.
}
\description{
A potentially more solid version to find the r.snippet
}
